Some improvements regarding const-correctness:
Signals of AbstractAspect and AbstractColumn are emitted from within undo
commands. Hence, changing the originator in reaction to these signals is almost
certainly a bug. In other words, the pointer to the Aspect/Column being changed
is conceptually const.

Likewise, having a filter change its inputs is dangerous and should be
discouraged. SimpleMappingFilter needs some rethinking; if possible in the
context of an actual use case.