Re-activate builtin Origin OPJ import using liborigin2.
Patch by Miquel Garriga merged from import_origin branch.

The feature is disabled by default in order to avoid complicating the build
process and because it's (in its current form) not tested thoroughly.
Nevertheless, including it as an option in the next release will give it some
more exposure (and certainly isn't a regression).